data-engineering/requirements.txt
psycopg2-binary==2.9.9
boto3
pyarrow
pyspark
neo4j
pandas